Created and modified files - added execute_process to cmake to build documentation during code build. implement basic parsing methods

+++ src/environment.c
@@ -1,6 +1,6 @@
 #include <unistd.h> // working dir, symbolic constants
 #include <stddef.h> // type aliases
-#include <stdlib.h> // environmental variables
+#include <stdlib.h> // environmental variables, malloc, free
 
 #include "environment.h"
 
@@ -13,12 +13,14 @@ void environment_init(Environment* env)
 	env->PATH = getenv("PATH") ;
 	env->USER = getenv("USER") ;
 	{
-		long n = pathconf(".", _PC_PATH_MAX) ;
+		size_t n = pathconf(".", _PC_PATH_MAX) ;
 		env->WORKING_DIRECTORY = malloc(n * sizeof(char)) ;
 		for(size_t i = 0 ; i < (size_t)n ; ++i) env->WORKING_DIRECTORY[i] = '\0' ;
 		getcwd(env->WORKING_DIRECTORY, n) ;
 	}
 }
+
+//void change_directory(Environment* env, char* 
 
 void environment_fini(Environment* env)
 {
